requirements.txt
aiohttp
autogen-agentchat==0.4.7
autogen-core==0.4.7
autogen-ext==0.4.7
fastmcp==1.0
flask
langchain-community
langchain-core
langchain-mcp-adapters>=0.0.1
langchain-openai
langchain_google_community
langgraph>=0.2.28
openai
pre-commit
psycopg2-binary
python-dotenv
pytz
Requests==2.32.3
sqlparse==0.4.4
streamlit
tabulate==0.9.0
tiktoken==0.8.0
